+++ eng/tools/rush-runner/tsconfig.json
@@ -5,5 +5,5 @@
     "target": "Es2021",
     "skipLibCheck": true
   },
-  "include": ["index.js", "test/**/*.js"]
+  "include": ["index.js", "src/**/*.js", "test/**/*.js"]
 }
